Rename KMessageWidget files to match our coding style
