+++ src/gui/KMessageWidget.h
